refactor: moved noop logger to ConductorLogger class
