+++ src/common/ConductorLogger.ts
@@ -61,3 +61,12 @@ export class DefaultLogger implements ConductorLogger {
     this.log("ERROR", ...args)
   }
 }
+
+export const noopLogger: ConductorLogger = {
+  //eslint-disable-next-line
+  debug: (...args: any) => {},
+  //eslint-disable-next-line
+  info: (...args: any) => {},
+  //eslint-disable-next-line
+  error: (...args: any) => {},
+};
